requirements.txt
fastapi
uvicorn
pillow
pytesseract
openai>=1.0.0
python-multipart
